Moved Tester to Widgest, Fixed Callback error
